+++ docs-source/docs/modules/cqrs/examples/shopping-cart-service-scala/src/main/protobuf/ShoppingCartService.proto
@@ -1,8 +1,7 @@
 syntax = "proto3";
 
 option java_multiple_files = true;
-option java_package = "sample.shoppingcart.proto";
-option java_outer_classname = "ShoppingCartProto";
+option java_package = "shopping.cart.proto";
 
 package shoppingcart;
