+++ docs-source/docs/modules/cqrs/examples/shopping-cart-service-scala/src/main/protobuf/ShoppingOrderService.proto
@@ -2,8 +2,7 @@
 syntax = "proto3";
 
 option java_multiple_files = true;
-option java_package = "sample.shoppingorder.proto";
-option java_outer_classname = "ShoppingOrderProto";
+option java_package = "shopping.order.proto";
 
 package shoppingorder;
