+++ docs-source/docs/modules/cqrs/examples/shopping-cart-service-scala/src/main/resources/application.conf
@@ -11,7 +11,7 @@ akka {
   # tag::cbor[]
   actor {
     serialization-bindings {
-      "sample.shoppingcart.CborSerializable" = jackson-cbor
+      "shopping.cart.CborSerializable" = jackson-cbor
     }
   }
   # end::cbor[]
