+++ docs-source/docs/modules/cqrs/examples/shopping-cart-service-scala/src/main/scala/shopping/cart/CborSerializable.scala
@@ -1,4 +1,4 @@
-package sample.shoppingcart
+package shopping.cart
 
 /**
  * Marker trait for serialization with Jackson CBOR.
