+++ docs-source/docs/modules/cqrs/examples/shopping-cart-service-scala/src/main/scala/shopping/cart/ItemPopularityRepository.scala
@@ -1,4 +1,4 @@
-package sample.shoppingcart
+package shopping.cart
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
