+++ docs-source/docs/modules/cqrs/examples/shopping-cart-service-scala/src/main/scala/shopping/cart/SendOrderProjection.scala
@@ -1,4 +1,4 @@
-package sample.shoppingcart
+package shopping.cart
 
 import akka.actor.typed.ActorSystem
 import akka.cluster.sharding.typed.ShardedDaemonProcessSettings
@@ -12,35 +12,34 @@ import akka.projection.eventsourced.EventEnvelope
 import akka.projection.eventsourced.scaladsl.EventSourcedProvider
 import akka.projection.scaladsl.AtLeastOnceProjection
 import akka.projection.scaladsl.SourceProvider
+import shopping.order.proto.ShoppingOrderService
 
-object ItemPopularityProjection {
-  def init(system: ActorSystem[_], repository: ItemPopularityRepository): Unit = {
-    ShardedDaemonProcess(system).init( // <1>
-      name = "ItemPopularityProjection",
+object SendOrderProjection {
+
+  def init(system: ActorSystem[_], orderService: ShoppingOrderService): Unit = {
+    ShardedDaemonProcess(system).init(
+      name = "SendOrderProjection",
       ShoppingCart.tags.size,
-      index => ProjectionBehavior(createProjectionFor(system, repository, index)),
+      index => ProjectionBehavior(createProjectionFor(system, orderService, index)),
       ShardedDaemonProcessSettings(system),
       Some(ProjectionBehavior.Stop))
   }
 
   private def createProjectionFor(
       system: ActorSystem[_],
-      repository: ItemPopularityRepository,
+      orderService: ShoppingOrderService,
       index: Int): AtLeastOnceProjection[Offset, EventEnvelope[ShoppingCart.Event]] = {
-    val tag = ShoppingCart.tags(index) // <2>
-    // tag::projection[]
-    val sourceProvider: SourceProvider[Offset, EventEnvelope[ShoppingCart.Event]] = // <3>
+    val tag = ShoppingCart.tags(index)
+    val sourceProvider: SourceProvider[Offset, EventEnvelope[ShoppingCart.Event]] =
       EventSourcedProvider.eventsByTag[ShoppingCart.Event](
         system = system,
-        readJournalPluginId = CassandraReadJournal.Identifier, // <4>
+        readJournalPluginId = CassandraReadJournal.Identifier,
         tag = tag)
 
-    CassandraProjection.atLeastOnce( // <5>
-      projectionId = ProjectionId("item-popularity", tag),
+    CassandraProjection.atLeastOnce(
+      projectionId = ProjectionId("orders", tag),
       sourceProvider,
-      handler = () => new ItemPopularityProjectionHandler(tag, system, repository)
-    ) // <6>
-    // end::projection[]
+      handler = () => new SendOrderProjectionHandler(system, orderService))
   }
 
 }
