+++ docs-source/docs/modules/cqrs/examples/shopping-cart-service-scala/src/main/scala/shopping/cart/ShoppingCartServer.scala
@@ -1,4 +1,4 @@
-package sample.shoppingorder
+package shopping.cart
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
@@ -13,17 +13,22 @@ import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.HttpRequest
 import akka.http.scaladsl.model.HttpResponse
 
-object ShoppingOrderServer {
+object ShoppingCartServer {
 
-  def start(interface: String, port: Int, system: ActorSystem[_]): Unit = {
+  def start(
+      interface: String,
+      port: Int,
+      system: ActorSystem[_],
+      itemPopularityRepository: ItemPopularityRepository): Unit = {
     implicit val sys: ActorSystem[_] = system
+    // tag::start-grpc[]
     implicit val ec: ExecutionContext = system.executionContext
 
     val service: HttpRequest => Future[HttpResponse] =
       ServiceHandler.concatOrNotFound(
-        proto.ShoppingOrderServiceHandler.partial(new ShoppingOrderServiceImpl),
+        proto.ShoppingCartServiceHandler.partial(new ShoppingCartServiceImpl(itemPopularityRepository)),
         // ServerReflection enabled to support grpcurl without import-path and proto parameters
-        ServerReflection.partial(List(proto.ShoppingOrderService)))
+        ServerReflection.partial(List(proto.ShoppingCartService)))
 
     val bound =
       Http().newServerAt(interface, port).bind(service).map(_.addToCoordinatedShutdown(3.seconds))
@@ -31,10 +36,11 @@ object ShoppingOrderServer {
     bound.onComplete {
       case Success(binding) =>
         val address = binding.localAddress
-        system.log.info("Shopping order at gRPC server {}:{}", address.getHostString, address.getPort)
+        system.log.info("Shopping online at gRPC server {}:{}", address.getHostString, address.getPort)
       case Failure(ex) =>
         system.log.error("Failed to bind gRPC endpoint, terminating system", ex)
         system.terminate()
+      // end::start-grpc[]
     }
   }
