+++ docs-source/docs/modules/cqrs/examples/shopping-order-service-scala/src/main/protobuf/ShoppingOrderService.proto
@@ -1,8 +1,7 @@
 syntax = "proto3";
 
 option java_multiple_files = true;
-option java_package = "sample.shoppingorder.proto";
-option java_outer_classname = "ShoppingOrderProto";
+option java_package = "shopping.order.proto";
 
 package shoppingorder;
